requirements.txt
requests
pyyaml
cerberus